Making stuff proper

+++ outputs/GranularSynth/GSReverbDelayHandler.h
@@ -18,7 +18,7 @@ public:
     GSReverbDelayHandler(GranularSyntheziser* GS);
     bool sendData ();
     void setParameters(std::vector<std::string> ParameterList);
-    const int GS_MAX_DELAY = 200;
+    const float GS_MAX_DELAY = 200;
 private:
     
 };
